Add impl-dependant param
Also:
  * Unify handling of internal types in the abstract base.
